[#2283] Change middleware to work around load balancer IP issue
